Begin MVC system
