+++ src/Abstract/AbstractModel.java
@@ -1,0 +1,5 @@
+package Abstract;
+
+public class AbstractModel
+{
+}
