+++ src/Abstract/AbstractView.java
@@ -1,0 +1,11 @@
+package Abstract;
+
+public abstract class AbstractView<TModel extends AbstractModel>
+{
+    protected TModel model;
+
+    protected AbstractView(TModel model)
+    {
+        this.model = model;
+    }
+}
